User CRUD
Updated Address file for userId so that it matches up with id field in User.java. Added hashcode and equals in User.java. Deleted RegistrationException.  UseService updated to have address to be saved in one-to-one didiectional relationship.   ALso working on login validation. UserController updated with getmappings and posMappings. Login.hTML updated to have username input field and not email input field. for validation check.  Placeholders added in the registers for email and phone fields. RegisterConfirmation button taken out not needed.