Searching for Arrivals
Disabled security to work on other aspects of the project.  I coded a view, controller, services, domain and repository for searching arrivals in a time period given.  The search is saved in a database.  I am hoping to perform a search and call the rest API to get all arrivals for an airport.