updated searchService
I wanted to have the user input airport, startdate and enddate.  These are queri API parameters for airport arrivals.  The search controller handles these parameters and sends them to the service.  The search service calls the api request with those parameters to get all arrivals for that airport.  startdate and enddate use epoch time. So I need to convert it.  The service needs a way to save all arrivals to json into java objects to be saves in repository.